pyproject.toml
[tool.poetry]
name = "ecoindex-cli"
version = "0.2.0"
description = "`ecoindex-cli` is a CLI tool that let you make ecoindex tests on given pages"
authors = ["Vincent Vatelot <vincent.vatelot@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "http://www.ecoindex.fr"
include = [
    "LICENSE",
]

[tool.poetry.scripts]
ecoindex-cli = "cli:app"

[tool.poetry.dependencies]
python = "^3.8"
ecoindex = "^0.5.0"
typer = "^0.3.2"
pandas = "^1.3.0"
Jinja2 = "^3.0.1"
matplotlib = "^3.4.1"
Scrapy = "^2.5.0"
click-spinner = "^0.1.10"
selenium = "^3.141.0"
pydantic = "^1.8.2"
chromedriver-py = "^92.0.4515"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = {version = "^21.7b0", allow-prereleases = true}
pytest-cov = "^2.12.1"
coverage-badge = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
